+++ src/ent/autoplay_results.h
@@ -43,7 +43,7 @@ void string_builder_add_winning_player_confidence(StringBuilder *sb,
                                                   double p1_total,
                                                   uint64_t total_games);
 void autoplay_results_set_write_buffer_size(AutoplayResults *autoplay_results,
-                                            int write_buffer_size);
+                                            size_t write_buffer_size);
 void autoplay_results_set_data_paths(AutoplayResults *autoplay_results,
                                      const char *data_paths);
 void autoplay_results_set_ld(AutoplayResults *autoplay_results,
