+++ src/ent/sim_results.h
@@ -42,14 +42,14 @@ void sim_results_destroy(SimResults *sim_results);
 int sim_results_get_number_of_plays(const SimResults *sim_results);
 int sim_results_get_max_plies(const SimResults *sim_results);
 int sim_results_get_node_count(const SimResults *sim_results);
-int sim_results_get_iteration_count(const SimResults *sim_results);
+uint64_t sim_results_get_iteration_count(const SimResults *sim_results);
 SimmedPlay *sim_results_get_simmed_play(const SimResults *sim_results,
                                         int index);
 SimmedPlay *sim_results_get_sorted_simmed_play(SimResults *sim_results,
                                                int index);
 BAIResult *sim_results_get_bai_result(const SimResults *sim_results);
 
-void sim_results_set_iteration_count(SimResults *sim_results, int count);
+void sim_results_set_iteration_count(SimResults *sim_results, uint64_t count);
 void sim_results_lock_simmed_plays(SimResults *sim_results);
 void sim_results_unlock_simmed_plays(SimResults *sim_results);
 bool sim_results_sort_plays_by_win_rate(SimResults *sim_results);
